Add routing (refactored accordingly)
