+++ .eslintrc
@@ -1,7 +1,7 @@
 {
-  "extends": ["react-app", "airbnb", "plugin:jsx-a11y/recommended"],
-  "rules": {
-    "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }]
-  },
-  "plugins": ["jsx-a11y"]
+    "extends": ["react-app", "airbnb", "plugin:jsx-a11y/recommended"],
+"rules": {
+  "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }]
+},
+"plugins": ["jsx-a11y"]
 }
