+++ package.json
@@ -4,10 +4,11 @@
   "private": true,
   "dependencies": {
     "bootstrap": "^4.1.3",
-    "eslint": "^5.10.0",
+    "prop-types": "^15.6.2",
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
     "react-redux": "^5.1.1",
+    "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.1",
     "react-test-renderer": "^16.6.3",
     "reactstrap": "^6.5.0",
@@ -22,9 +23,6 @@
     "server": "nodemon server.js --prefix server",
     "dev": "concurrently \"npm run start\" \"npm run server\" "
   },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
   "browserslist": [
     ">0.2%",
     "not dead",
@@ -33,7 +31,6 @@
   ],
   "proxy": "http://localhost:5000",
   "devDependencies": {
-    "concurrently": "^4.1.0",
-    "eslint-config-airbnb": "^17.1.0"
+    "concurrently": "^4.1.0"
   }
 }
