--- src/App.js
@@ -1,16 +1,0 @@
-import React, { Component } from 'react';
-import logo from './logo.svg';
-import RegistrationPage from './RegistrationPage'
-import './App.css';
-
-class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <RegistrationPage />
-      </div>
-    );
-  }
-}
-
-export default App;
